# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Saitynas API",
    "version": "v1"
  },
  "paths": {
    "/api/v1/Evaluations": {
      "get": {
        "tags": [
          "Evaluations"
        ],
        "parameters": [
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEvaluationDTOGetListDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Evaluations"
        ],
        "parameters": [
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEvaluationDTOGetObjectDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Evaluations/{id}": {
      "get": {
        "tags": [
          "Evaluations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEvaluationDTOGetObjectDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Evaluations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/EditEvaluationDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditEvaluationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditEvaluationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditEvaluationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEvaluationDTOGetObjectDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Evaluations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Specialists/{id}": {
      "delete": {
        "tags": [
          "Specialists"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Specialities": {
      "get": {
        "tags": [
          "Specialities"
        ],
        "parameters": [
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEnumDTOGetListDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Specialities"
        ],
        "parameters": [
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/EnumDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnumDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnumDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnumDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEnumDTOGetObjectDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Specialities/{id}": {
      "get": {
        "tags": [
          "Specialities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnumDTOGetObjectDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Specialities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/EnumDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnumDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnumDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnumDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEnumDTOGetObjectDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Specialities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1": {
      "get": {
        "tags": [
          "Welcome"
        ],
        "parameters": [
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Workplaces": {
      "get": {
        "tags": [
          "Workplaces"
        ],
        "parameters": [
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkplaceDTOGetListDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workplaces"
        ],
        "parameters": [
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/WorkplaceDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkplaceDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkplaceDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WorkplaceDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkplaceDTOGetObjectDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Workplaces/{id}": {
      "get": {
        "tags": [
          "Workplaces"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkplaceDTOGetObjectDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Workplaces"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/WorkplaceDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkplaceDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkplaceDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WorkplaceDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkplaceDTOGetObjectDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workplaces"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Api-Request",
            "in": "header",
            "description": "Api Request Header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EditEvaluationDTO": {
        "type": "object",
        "properties": {
          "evaluation": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EnumDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EnumDTOGetObjectDTO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/EnumDTO"
          }
        },
        "additionalProperties": false
      },
      "EvaluationDTO": {
        "type": "object",
        "properties": {
          "evaluation": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetEnumDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetEnumDTOGetListDTO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetEnumDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetEnumDTOGetObjectDTO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/GetEnumDTO"
          }
        },
        "additionalProperties": false
      },
      "GetEvaluationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "evaluation": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetEvaluationDTOGetListDTO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetEvaluationDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetEvaluationDTOGetObjectDTO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/GetEvaluationDTO"
          }
        },
        "additionalProperties": false
      },
      "GetWorkplaceDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetWorkplaceDTOGetListDTO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetWorkplaceDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetWorkplaceDTOGetObjectDTO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/GetWorkplaceDTO"
          }
        },
        "additionalProperties": false
      },
      "MessageData": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MessageDTO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/MessageData"
          }
        },
        "additionalProperties": false
      },
      "Meta": {
        "type": "object",
        "additionalProperties": false
      },
      "WorkplaceDTO": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}